let the library VERSION = the project's one (issue #490)
This is a cross-platform usability improvement.
On Unix platforms it is customary for library to have VERSION
and SOVERSION, where SOVERSION changes on major API changes
and VERSION is the same as project's version, so library users
always know what vesion this library belongs to just by name.
With this patch we have a proper libmimalloc.so.VERSION on Unix.